requirements/base.txt
# Django
Django==1.10.5

django-cors-headers==2.0

# Environmental var configuration
django-configurations==2.0

# Subdomain routing
django-subdomains==2.1.0

# Encrypted fields
django-fernet-fields==0.5.0

# Stripe client
stripe==1.48.1

# For the persistence stores
psycopg2==2.6.2
dj-database-url==0.4.2

# Models
django-model-utils==2.6.1

# Rest apis
djangorestframework==3.5.3
Markdown==2.6.7
django-filter==1.0.1

# Static files
whitenoise==3.2.3

# Media files
boto==2.45.0
django_unique_upload==0.2.1

# Time zones support
pytz==2016.10

# Python requests
requests==2.13.0

# IP finding
django-ipware==1.1.6

# For asynchronous queuing		
django-rq==0.9.4
